封装函数
